+++ modular_sdk/models/pynamodb_extension/base_model.py
@@ -25,7 +25,7 @@ from pynamodb.settings import OperationSettings
 
 from modular_sdk.commons.constants import MODULAR_SERVICE_MODE_ENV, \
     SERVICE_MODE_DOCKER, PARAM_MONGO_USER, PARAM_MONGO_PASSWORD, \
-    PARAM_MONGO_URL, PARAM_MONGO_DB_NAME
+    PARAM_MONGO_URL, PARAM_MONGO_DB_NAME, PARAM_MONGO_SRV
 from modular_sdk.commons.helpers import classproperty
 from modular_sdk.commons.log_helper import get_logger
 from modular_sdk.commons.time_helper import utc_iso
@@ -33,11 +33,13 @@ from modular_sdk.commons.time_helper import utc_iso
 _LOG = get_logger(__name__)
 
 
-def build_mongodb_uri(user: str, password: str, url: str) -> str:
+def build_mongodb_uri(user: str, password: str, url: str,
+                      srv: bool = False) -> str:
     """
     Just makes the right formatting
     """
-    return f'mongodb://{user}:{password}@{url}/'
+
+    return f'mongodb{"+srv" if srv else ''}://{user}:{password}@{url}/'
 
 
 class M3BooleanAttribute(BooleanAttribute):
@@ -88,6 +90,7 @@ class DynamicAttribute(Attribute):
         bool: BooleanAttribute,
         bytes: JSONAttribute  # todo, BinaryAttribute would fit better but
     }
+
     # ... but this class has a bug -> it does not perform data
     # deserialization. The raw value from DB is returned all the time.
     # For Unicode, Number, List, Bool it is more or less acceptable, and
@@ -194,9 +197,12 @@ class ModularMongoDBHandlerMixin(ABCMongoDBHandlerMixin):
             password = os.environ.get(PARAM_MONGO_PASSWORD)
             url = os.environ.get(PARAM_MONGO_URL)
             db = os.environ.get(PARAM_MONGO_DB_NAME)
+            srv = os.environ.get(PARAM_MONGO_SRV)
+            srv = True if srv and srv.lower() in ('true', 't', 'y', 'yes') \
+                else False
             cls._mongodb = PynamoDBToPyMongoAdapter(
                 mongodb_connection=MongoDBConnection(
-                    build_mongodb_uri(user, password, url), db
+                    build_mongodb_uri(user, password, url, srv), db
                 )
             )
         return cls._mongodb