+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [7.0.0] - 2024-12-18
-- rewrite Pynamo & Mongo abstraction to support all attributes and functionality
-- Removed some existing variables, functions, classes and modules:
+- Rewritten Pynamo & Mongo abstraction to support all attributes and functionality
+- Updated `pymongo` version to `pymongo~=4.10.1`
+- Removed variables, functions, classes and modules:
   - `modular_sdk.commons.DynamoDBJsonSerializer`
   - `modular_sdk.commons.build_response`
   - `modular_sdk.commons.RESPONSE_*`
@@ -14,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `modular_sdk.commons.error_helper`
   - `modular_sdk.commons.deep_pop`
   - `modular_sdk.helpers`
+  - `modular_sdk.models.pynamodb_extension`
+  - `modular_sdk.connections.mongodb_connection`
+- Added:
+  - `modular_sdk.models.pynamongo.index_creator.IndexCreator`
 
 ## [6.3.0] - 2024-11-26
 - Fix Vault parameters encoder to comply with Maestro format