+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `get_json()` is not deprecated and will be removed. Use pynamongo.convertors.instance_to_dict or write your serializer
   - `mongo_id` attribute was replaced with `__mongo_id__`
 - Updated `pymongo` version to `pymongo~=4.10.1`
+- Removed `dynamodb-json~=1.4.2` package
 - Removed variables, functions, classes and modules:
   - `modular_sdk.commons.DynamoDBJsonSerializer`
   - `modular_sdk.commons.build_response`