+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.4.0] - 2025-02-05
+- Support mongodb SRV connection strings with `modular_mongo_srv` env.
+
 ## [6.3.0] - 2024-11-26
 - Fix Vault parameters encoder to comply with Maestro format
 - Update `get_signed_headers` to support conditional handling of `http` and